pyproject.toml
[tool.poetry]
name = "aoc"
version = "0.1.0"
description = "Python template for Advent of Code event"
authors = ["Leonardo Gago <ljgago@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
colorama = "^0.4.4"
python-dotenv = "^0.19.1"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[tool.pytest.ini_options]
addopts = "-v"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
